Since we are working with floats, output in float too

+++ src/main/java/org/uu/nl/embedding/opt/IOptimizer.java
@@ -7,5 +7,5 @@ public interface IOptimizer {
 	Optimum optimize() throws OptimizationFailedException;
 	String getName();
 	OptimizeJob createJob(int id, int iteration);
-	double[] extractResult();
+	float[] extractResult();
 }
